RCS_Tutorial
The attached Matlab code runs through and explains all of the necessary commands needed to create a unique radar cross section for an object in STK using the object model. A brief explanation is given in the code to clarify why certain commands are called in the particular order. This essentially explains how the code mirrors the GUI, since some of the process is not intuitive.